redux code restructured
